flow_analysis/requirements.txt
# API and Data Processing
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2023.3

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Monitoring and Metrics
psutil>=5.9.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.3.0
pytest-mock>=3.11.0

# Error Handling and Retries
backoff>=2.2.1
tenacity>=8.2.0

# Logging and Utilities
python-json-logger>=2.0.7
python-dotenv>=1.0.0

# Development Tools
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0 